- rendering: allow to specify the rendering order of shaders through draw queues - rendering: distinguish opaque shaders from transparent ones - rendering: allow to restrict rendering of objects to specific cameras through layers - rendering: replace the local to view matrix components with a local to world component - rendering: allow to override the default rendering pipeline with a custom scene method

+++ include/Material.hpp
@@ -9,7 +9,7 @@
 
 namespace Age::Gfx
 {
-using MaterialId = std::uint32_t;
+using MaterialId = std::uint16_t;
 
 struct Material
 {
